requirements.txt
flask
requests
PyJWT[crypto] ~= 1.7.1
